Moved WSHandler to its own file, added ws communication between /train and /receive_image for message display.
